+++ relayer/spy_relayer/.env.tilt.relayer
@@ -1,10 +1,10 @@
-SUPPORTED_CHAINS=[  {    "chainId": 1,    "chainName": "Solana",  "nativeCurrencySymbol": "SOL",  "nodeUrl": "http://solana-devnet:8899",    "tokenBridgeAddress": "0x0290FB167208Af455bB137780163b7B7a9a10C16",    "bridgeAddress": "Bridge1p5gheXUvJ6jGWGeCsgPKgnE3YgdGKRVCMY9o",    "walletPrivateKey": [      [        14, 173, 153, 4, 176, 224, 201, 111, 32, 237, 183, 185, 159, 247, 22,        161, 89, 84, 215, 209, 212, 137, 10, 92, 157, 49, 29, 192, 101, 164,        152, 70, 87, 65, 8, 174, 214, 157, 175, 126, 98, 90, 54, 24, 100, 177,        247, 77, 19, 112, 47, 44, 165, 109, 233, 102, 14, 86, 109, 29, 134, 145,        132, 141      ]    ],    "wrappedAsset": "So11111111111111111111111111111111111111112"  },  {    "chainId": 2,    "chainName": "Ethereum",  "nativeCurrencySymbol": "ETH",  "nodeUrl": "http://eth-devnet:8545",    "tokenBridgeAddress": "0x0290FB167208Af455bB137780163b7B7a9a10C16",    "walletPrivateKey": [      "0x4f3edf983ac636a65a842ce7c78d9aa706d3b113bce9c46f30d7d21715b23b1d"    ],    "wrappedAsset": "0xDDb64fE46a91D46ee29420539FC25FD07c5FEa3E"  },  {    "chainId": 3,    "chainName": "Terra",  "nativeCurrencySymbol": "LUNA",  "nodeUrl": "http://terra-terrad:1317",    "tokenBridgeAddress": "terra10pyejy66429refv3g35g2t7am0was7ya7kz2a4",    "walletPrivateKey": [      "notice oak worry limit wrap speak medal online prefer cluster roof addict wrist behave treat actual wasp year salad speed social layer crew genius"    ],    "terraName": "localterra",    "terraChainId": "columbus-5",    "terraCoin": "uluna",    "terraGasPriceUrl": "http://terra-fcd:3060/v1/txs/gas_prices"  },  {    "chainId": 4,    "chainName": "Binance Smart Chain",  "nativeCurrencySymbol": "BNB",  "nodeUrl": "http://eth-devnet2:8545",    "tokenBridgeAddress": "0x0290FB167208Af455bB137780163b7B7a9a10C16",    "walletPrivateKey": [      "0x4f3edf983ac636a65a842ce7c78d9aa706d3b113bce9c46f30d7d21715b23b1d"    ],    "wrappedAsset": "0xDDb64fE46a91D46ee29420539FC25FD07c5FEa3E"  }]
-REDIS_HOST= redis
+SUPPORTED_CHAINS=[{"chainId":1,"chainName":"Solana","nodeUrl":"http://solana-devnet:8899","nativeCurrencySymbol":"SOL","tokenBridgeAddress":"B6RHG3mfcckmrYN1UhmJzyS1XX3fZKbkeUcpJe9Sy3FE","bridgeAddress":"Bridge1p5gheXUvJ6jGWGeCsgPKgnE3YgdGKRVCMY9o","wrappedAsset":"So11111111111111111111111111111111111111112"},{"chainId":2,"chainName":"ETH","nativeCurrencySymbol":"ETH","nodeUrl":"http://eth-devnet:8545","tokenBridgeAddress":"0x0290FB167208Af455bB137780163b7B7a9a10C16","wrappedAsset":"0xDDb64fE46a91D46ee29420539FC25FD07c5FEa3E"},{"chainId":3,"chainName":"Terra","isTerraClassic":true,"nativeCurrencySymbol":"LUNA","nodeUrl":"http://terra-terrad:1317","tokenBridgeAddress":"terra10pyejy66429refv3g35g2t7am0was7ya7kz2a4","terraName":"localterra","terraChainId":"localterra","terraCoin":"uluna","terraGasPriceUrl":"http://terra-fcd:3060/v1/txs/gas_prices"},{"chainId":4,"chainName":"BSC","nativeCurrencySymbol":"BNB","nodeUrl":"http://eth-devnet2:8545","tokenBridgeAddress":"0x0290FB167208Af455bB137780163b7B7a9a10C16","wrappedAsset":"0xDDb64fE46a91D46ee29420539FC25FD07c5FEa3E"}]
+REDIS_HOST=redis
 REDIS_PORT=6379
 PROM_PORT=8083
 READINESS_PORT=2000
-CLEAR_REDIS_ON_INIT=false
+CLEAR_REDIS_ON_INIT=true
 DEMOTE_WORKING_ON_INIT=true
 LOG_LEVEL=debug
-SUPPORTED_TOKENS=[{"chainId":1,"address":"So11111111111111111111111111111111111111112"}, {"chainId":2,"address":"0xDDb64fE46a91D46ee29420539FC25FD07c5FEa3E"}, {"chainId":3,"address":"uluna"}, {"chainId":4,"address":"0xDDb64fE46a91D46ee29420539FC25FD07c5FEa3E"}]
-PRIVATE_KEYS=[  {    "chainId": 1,    "privateKeys": [      [        14, 173, 153, 4, 176, 224, 201, 111, 32, 237, 183, 185, 159, 247, 22,        161, 89, 84, 215, 209, 212, 137, 10, 92, 157, 49, 29, 192, 101, 164,        152, 70, 87, 65, 8, 174, 214, 157, 175, 126, 98, 90, 54, 24, 100, 177,        247, 77, 19, 112, 47, 44, 165, 109, 233, 102, 14, 86, 109, 29, 134, 145,        132, 141      ]    ]  },  {    "chainId": 2,    "privateKeys": [      "0x4f3edf983ac636a65a842ce7c78d9aa706d3b113bce9c46f30d7d21715b23b1d"    ]  },  {    "chainId": 3,    "privateKeys": [      "notice oak worry limit wrap speak medal online prefer cluster roof addict wrist behave treat actual wasp year salad speed social layer crew genius"    ]  },  {    "chainId": 4,    "privateKeys": [      "0x4f3edf983ac636a65a842ce7c78d9aa706d3b113bce9c46f30d7d21715b23b1d"    ]  }]
+SUPPORTED_TOKENS=[{"chainId":1,"address":"So11111111111111111111111111111111111111112"},{"chainId":1,"address":"2WDq7wSs9zYrpx2kbHDA4RUTRch2CCTP6ZWaH4GNfnQQ"},{"chainId":2,"address":"0xDDb64fE46a91D46ee29420539FC25FD07c5FEa3E"},{"chainId":2,"address":"0x2D8BE6BF0baA74e0A907016679CaE9190e80dD0A"},{"chainId":3,"address":"uluna"},{"chainId":4,"address":"0xDDb64fE46a91D46ee29420539FC25FD07c5FEa3E"}]
+PRIVATE_KEYS=[{"chainId":1,"privateKeys":[[14,173,153,4,176,224,201,111,32,237,183,185,159,247,22,161,89,84,215,209,212,137,10,92,157,49,29,192,101,164,152,70,87,65,8,174,214,157,175,126,98,90,54,24,100,177,247,77,19,112,47,44,165,109,233,102,14,86,109,29,134,145,132,141]]},{"chainId":2,"privateKeys":["0x4f3edf983ac636a65a842ce7c78d9aa706d3b113bce9c46f30d7d21715b23b1d"]},{"chainId":3,"privateKeys":["notice oak worry limit wrap speak medal online prefer cluster roof addict wrist behave treat actual wasp year salad speed social layer crew genius"]},{"chainId":4,"privateKeys":["0x4f3edf983ac636a65a842ce7c78d9aa706d3b113bce9c46f30d7d21715b23b1d"]}]
