+++ relayer/spy_relayer/.vscode/launch.json
@@ -1,0 +1,18 @@
+{
+  // Use IntelliSense to learn about possible attributes.
+  // Hover to view descriptions of existing attributes.
+  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "type": "node",
+      "request": "launch",
+      "name": "spy_relay",
+      "skipFiles": ["<node_internals>/**"],
+      "program": "${workspaceFolder}/lib/main.js",
+      "outFiles": ["${workspaceFolder}/lib/**/*.js"],
+      "preLaunchTask": "npm: build",
+      "outputCapture": "std"
+    }
+  ]
+}
