--- relayer/spy_relayer/QA.md
@@ -1,3 +1,0 @@
-Test Scenarios:
-
-- Should survive any container restarting: Spy, Listener, Redis, or Relayer.
