+++ relayer/spy_relayer/src/backends/default/index.ts
@@ -1,0 +1,11 @@
+import { Backend, Relayer, Listener } from "../definitions";
+import { TokenBridgeListener } from "./listener";
+import { TokenBridgeRelayer } from "./relayer";
+
+/** Payload version 1 token bridge listener and relayer backend */
+const backend: Backend = {
+  relayer: new TokenBridgeRelayer(),
+  listener: new TokenBridgeListener(),
+};
+
+export default backend;
