+++ relayer/spy_relayer/tsconfig.json
@@ -12,7 +12,8 @@
     "noFallthroughCasesInSwitch": true,
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "downlevelIteration": true
+    "downlevelIteration": true,
+    "sourceMap": true
   },
   "include": ["src"],
   "exclude": ["node_modules", "**/__tests__/*"]
